requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
gunicorn
psycopg2-binary
celery
redis
django-environ
django-filter>=23.3
django-celery-beat>=2.5
pytest
pytest-django
pytest-cov
factory_boy
docker
mock
faker

